rename agents list and entries list table
