+++ tornjak-frontend/src/tables/entries-list-table.js
